code cleanup, removing neccesary (un)boxing
git-svn-id: https://svn.kenai.com/svn/bsaf~main/trunk@107 39130dca-0e0f-c9f4-f3e0-9e7262a0d940

+++ bsaf/src/main/java/org/jdesktop/application/DefaultInputBlocker.java
@@ -342,7 +342,7 @@ final class DefaultInputBlocker extends Task.InputBlocker {
         if ((delay == null) && (taskResourceMap != null)) {
             delay = taskResourceMap.getInteger(key);
         }
-        return (delay == null) ? 0 : delay.intValue();
+        return (delay == null) ? 0 : delay;
     }
 
     private void showBlockingDialog(boolean f) {
